VM tests refactoring
