+++ evm-tests/jsontests/Cargo.toml
@@ -31,6 +31,7 @@ parity-bytes = "0.1"
 hex-literal = "0.4"
 
 ethjson = { path = "../ethjson", features = ["test-helpers"] }
+ethereum = "0.15"
 ethcore-builtin = { path = "../ethcore-builtin" }
 
 [features]
