+++ evm-tests/jsontests/src/state.rs
@@ -1,35 +1,20 @@
-use crate::execution_results::{FailedTestDetails, TestExecutionResult};
-use crate::state_dump::StateTestsDump;
-use crate::types::spec::Spec;
+use crate::execution_results::TestExecutionResult;
+use crate::types::Spec;
 use crate::types::StateTestCase;
-use crate::utils::eip_4844;
 use crate::utils::transaction::InvalidTxReason;
-use aurora_evm::backend::{ApplyBackend, MemoryAccount, MemoryBackend, MemoryVicinity};
-use aurora_evm::executor::stack::{
-    Authorization, MemoryStackState, PrecompileFailure, PrecompileFn, PrecompileOutput,
-    StackExecutor, StackSubstateMetadata,
-};
-use aurora_evm::utils::U64_MAX;
-use aurora_evm::{Config, Context, ExitError, ExitReason, ExitSucceed};
+use aurora_evm::backend::MemoryVicinity;
 
+use crate::config::TestConfig;
 use libsecp256k1::SecretKey;
 use primitive_types::{H160, H256, U256};
 use serde::Deserialize;
 use sha3::{Digest, Keccak256};
-use std::collections::BTreeMap;
-use std::path::PathBuf;
 use std::str::FromStr;
-use std::sync::{Arc, LazyLock};
 
 #[derive(Deserialize, Debug)]
 pub struct Test(ethjson::test_helpers::state::State);
 
 impl Test {
-    #[must_use]
-    pub fn unwrap_to_pre_state(&self) -> BTreeMap<H160, MemoryAccount> {
-        crate::utils::unwrap_to_state(&self.0.pre_state)
-    }
-
     #[allow(clippy::missing_panics_doc)]
     #[must_use]
     pub fn unwrap_caller_secret_key(&self) -> H256 {
@@ -72,13 +57,13 @@ impl Test {
     /// Panics occurs for invalid transaction data.
     pub fn unwrap_to_vicinity(
         &self,
-        spec: &ForkSpec,
+        spec: &Spec,
         blob_gas_price: Option<u128>,
     ) -> Result<MemoryVicinity, InvalidTxReason> {
         let block_base_fee_per_gas = self.0.env.block_base_fee_per_gas.0;
         let tx = &self.0.transaction;
         // Validation for EIP-1559 that was introduced in London hard fork
-        let gas_price = if *spec >= ForkSpec::London {
+        let gas_price = if *spec >= Spec::London {
             tx.gas_price.or(tx.max_fee_per_gas).unwrap_or_default().0
         } else {
             if tx.max_fee_per_gas.is_some() {
@@ -106,7 +91,7 @@ impl Test {
             return Err(InvalidTxReason::GasPriceLessThenBlockBaseFee);
         }
 
-        let block_randomness = if spec.is_eth2() {
+        let block_randomness = if *spec > Spec::Berlin {
             self.0.env.random.map(|r| {
                 // Convert between U256 and H256. U256 is in little-endian but since H256 is just
                 // a string-like byte array, it's big endian (MSB is the first element of the array).
@@ -141,476 +126,6 @@ impl Test {
     }
 }
 
-#[derive(Debug, Clone)]
-struct Precompile;
-
-type LazyPrecompiles = LazyLock<BTreeMap<H160, Precompile>>;
-static ISTANBUL_BUILTINS: LazyPrecompiles = LazyLock::new(istanbul_builtins);
-static BERLIN_BUILTINS: LazyPrecompiles = LazyLock::new(berlin_builtins);
-static CANCUN_BUILTINS: LazyPrecompiles = LazyLock::new(cancun_builtins);
-static PRAGUE_BUILTINS: LazyPrecompiles = LazyLock::new(prague_builtins);
-
-macro_rules! precompile_entry {
-    ($map:expr, $builtins:expr, $index:expr) => {
-        let x: PrecompileFn =
-            |input: &[u8], gas_limit: Option<u64>, _context: &Context, _is_static: bool| {
-                let builtin = $builtins.get(&H160::from_low_u64_be($index)).unwrap();
-                Self::exec_as_precompile(builtin, input, gas_limit)
-            };
-        $map.insert(H160::from_low_u64_be($index), x);
-    };
-}
-
-pub struct JsonPrecompile;
-
-impl JsonPrecompile {
-    #[allow(clippy::match_same_arms)]
-    #[must_use]
-    pub fn precompile(spec: &ForkSpec) -> Option<BTreeMap<H160, PrecompileFn>> {
-        match spec {
-            ForkSpec::Istanbul => {
-                let mut map = BTreeMap::new();
-                precompile_entry!(map, ISTANBUL_BUILTINS, 1);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 2);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 3);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 4);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 5);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 6);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 7);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 8);
-                precompile_entry!(map, ISTANBUL_BUILTINS, 9);
-                Some(map)
-            }
-            ForkSpec::Berlin => {
-                let mut map = BTreeMap::new();
-                precompile_entry!(map, BERLIN_BUILTINS, 1);
-                precompile_entry!(map, BERLIN_BUILTINS, 2);
-                precompile_entry!(map, BERLIN_BUILTINS, 3);
-                precompile_entry!(map, BERLIN_BUILTINS, 4);
-                precompile_entry!(map, BERLIN_BUILTINS, 5);
-                precompile_entry!(map, BERLIN_BUILTINS, 6);
-                precompile_entry!(map, BERLIN_BUILTINS, 7);
-                precompile_entry!(map, BERLIN_BUILTINS, 8);
-                precompile_entry!(map, BERLIN_BUILTINS, 9);
-                Some(map)
-            }
-            // precompiles for London and Berlin are the same
-            ForkSpec::London => Self::precompile(&ForkSpec::Berlin),
-            // precompiles for Merge and Berlin are the same
-            ForkSpec::Merge => Self::precompile(&ForkSpec::Berlin),
-            // precompiles for Paris and Berlin are the same
-            ForkSpec::Paris => Self::precompile(&ForkSpec::Berlin),
-            // precompiles for Shanghai and Berlin are the same
-            ForkSpec::Shanghai => Self::precompile(&ForkSpec::Berlin),
-            ForkSpec::Cancun => {
-                let mut map = BTreeMap::new();
-                precompile_entry!(map, CANCUN_BUILTINS, 1);
-                precompile_entry!(map, CANCUN_BUILTINS, 2);
-                precompile_entry!(map, CANCUN_BUILTINS, 3);
-                precompile_entry!(map, CANCUN_BUILTINS, 4);
-                precompile_entry!(map, CANCUN_BUILTINS, 5);
-                precompile_entry!(map, CANCUN_BUILTINS, 6);
-                precompile_entry!(map, CANCUN_BUILTINS, 7);
-                precompile_entry!(map, CANCUN_BUILTINS, 8);
-                precompile_entry!(map, CANCUN_BUILTINS, 9);
-                precompile_entry!(map, CANCUN_BUILTINS, 0xA);
-                Some(map)
-            }
-            ForkSpec::Prague => {
-                let mut map = BTreeMap::new();
-                precompile_entry!(map, PRAGUE_BUILTINS, 1);
-                precompile_entry!(map, PRAGUE_BUILTINS, 2);
-                precompile_entry!(map, PRAGUE_BUILTINS, 3);
-                precompile_entry!(map, PRAGUE_BUILTINS, 4);
-                precompile_entry!(map, PRAGUE_BUILTINS, 5);
-                precompile_entry!(map, PRAGUE_BUILTINS, 6);
-                precompile_entry!(map, PRAGUE_BUILTINS, 7);
-                precompile_entry!(map, PRAGUE_BUILTINS, 8);
-                precompile_entry!(map, PRAGUE_BUILTINS, 9);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0A);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0B);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0C);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0D);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0E);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x0F);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x10);
-                precompile_entry!(map, PRAGUE_BUILTINS, 0x11);
-                Some(map)
-            }
-            _ => None,
-        }
-    }
-
-    fn exec_as_precompile(
-        builtin: &ethcore_builtin::Builtin,
-        input: &[u8],
-        gas_limit: Option<u64>,
-    ) -> Result<(PrecompileOutput, u64), PrecompileFailure> {
-        let cost = builtin.cost(input, 0);
-
-        if let Some(target_gas) = gas_limit {
-            if cost > U64_MAX || target_gas < cost.as_u64() {
-                return Err(PrecompileFailure::Error {
-                    exit_status: ExitError::OutOfGas,
-                });
-            }
-        }
-
-        let mut output = Vec::new();
-        match builtin.execute(input, &mut parity_bytes::BytesRef::Flexible(&mut output)) {
-            Ok(()) => Ok((
-                PrecompileOutput {
-                    exit_status: ExitSucceed::Stopped,
-                    output,
-                },
-                cost.as_u64(),
-            )),
-            Err(e) => Err(PrecompileFailure::Error {
-                exit_status: ExitError::Other(e.into()),
-            }),
-        }
-    }
-}
-
-#[allow(clippy::too_many_lines)]
-fn istanbul_builtins() -> BTreeMap<H160, Precompile> {
-    use ethjson::spec::builtin::{BuiltinCompat, Linear, Modexp, PricingCompat};
-
-    let builtins: BTreeMap<Address, BuiltinCompat> = BTreeMap::from([
-        (
-            Address(H160::from_low_u64_be(1)),
-            BuiltinCompat {
-                name: "ecrecover".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                    base: 3000,
-                    word: 0,
-                })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(2)),
-            BuiltinCompat {
-                name: "sha256".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 60, word: 12 })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(3)),
-            BuiltinCompat {
-                name: "ripemd160".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                    base: 600,
-                    word: 120,
-                })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(4)),
-            BuiltinCompat {
-                name: "identity".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 15, word: 3 })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(5)),
-            BuiltinCompat {
-                name: "modexp".to_string(),
-                pricing: PricingCompat::Single(Pricing::Modexp(Modexp {
-                    divisor: 20,
-                    is_eip_2565: false,
-                })),
-                activate_at: Some(Uint(U256::zero())),
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(6)),
-            BuiltinCompat {
-                name: "alt_bn128_add".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128ConstOperations(AltBn128ConstOperations {
-                            price: 150,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(7)),
-            BuiltinCompat {
-                name: "alt_bn128_mul".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128ConstOperations(AltBn128ConstOperations {
-                            price: 6000,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(8)),
-            BuiltinCompat {
-                name: "alt_bn128_pairing".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128Pairing(AltBn128Pairing {
-                            base: 45000,
-                            pair: 34000,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(9)),
-            BuiltinCompat {
-                name: "blake2_f".to_string(),
-                pricing: PricingCompat::Single(Pricing::Blake2F { gas_per_round: 1 }),
-                activate_at: Some(Uint(U256::zero())),
-            },
-        ),
-    ]);
-    builtins
-        .into_iter()
-        .map(|(address, builtin)| (address.into(), Precompile))
-        .collect()
-}
-
-#[allow(clippy::too_many_lines)]
-fn berlin_builtins() -> BTreeMap<H160, Precompile> {
-    let builtins: BTreeMap<Address, BuiltinCompat> = BTreeMap::from([
-        (
-            Address(H160::from_low_u64_be(1)),
-            BuiltinCompat {
-                name: "ecrecover".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                    base: 3000,
-                    word: 0,
-                })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(2)),
-            BuiltinCompat {
-                name: "sha256".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 60, word: 12 })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(3)),
-            BuiltinCompat {
-                name: "ripemd160".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                    base: 600,
-                    word: 120,
-                })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(4)),
-            BuiltinCompat {
-                name: "identity".to_string(),
-                pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 15, word: 3 })),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(5)),
-            BuiltinCompat {
-                name: "modexp".to_string(),
-                pricing: PricingCompat::Single(Pricing::Modexp(Modexp {
-                    divisor: 3,
-                    is_eip_2565: true,
-                })),
-                activate_at: Some(Uint(U256::zero())),
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(6)),
-            BuiltinCompat {
-                name: "alt_bn128_add".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128ConstOperations(AltBn128ConstOperations {
-                            price: 150,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(7)),
-            BuiltinCompat {
-                name: "alt_bn128_mul".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128ConstOperations(AltBn128ConstOperations {
-                            price: 6000,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(8)),
-            BuiltinCompat {
-                name: "alt_bn128_pairing".to_string(),
-                pricing: PricingCompat::Multi(BTreeMap::from([(
-                    Uint(U256::zero()),
-                    PricingAt {
-                        info: Some("EIP 1108 transition".to_string()),
-                        price: Pricing::AltBn128Pairing(AltBn128Pairing {
-                            base: 45000,
-                            pair: 34000,
-                        }),
-                    },
-                )])),
-                activate_at: None,
-            },
-        ),
-        (
-            Address(H160::from_low_u64_be(9)),
-            BuiltinCompat {
-                name: "blake2_f".to_string(),
-                pricing: PricingCompat::Single(Pricing::Blake2F { gas_per_round: 1 }),
-                activate_at: Some(Uint(U256::zero())),
-            },
-        ),
-    ]);
-    builtins
-        .into_iter()
-        .map(|(address, builtin)| {
-            (
-                address.into(),
-                ethjson::spec::Builtin::from(builtin).try_into().unwrap(),
-            )
-        })
-        .collect()
-}
-
-fn cancun_builtins() -> BTreeMap<H160, Precompile> {
-    let mut builtins = berlin_builtins();
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xA)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "kzg".to_string(),
-            pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                base: 50_000,
-                word: 0,
-            })),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins
-}
-
-fn prague_builtins() -> BTreeMap<H160, Precompile> {
-    let mut builtins = cancun_builtins();
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xB)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_g1_add".to_string(),
-            pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 375, word: 0 })),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xC)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_g1_mul".to_string(),
-            pricing: PricingCompat::Single(Pricing::Bls12G1Mul),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xD)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_g2_add".to_string(),
-            pricing: PricingCompat::Single(Pricing::Linear(Linear { base: 600, word: 0 })),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xE)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_g2_mul".to_string(),
-            pricing: PricingCompat::Single(Pricing::Bls12G2Mul),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0xF)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_pairing".to_string(),
-            pricing: PricingCompat::Single(Pricing::Bls12Pairing),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0x10)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_fp_to_g1".to_string(),
-            pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                base: 5_500,
-                word: 0,
-            })),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-    builtins.insert(
-        Address(H160::from_low_u64_be(0x11)).into(),
-        ethjson::spec::Builtin::from(BuiltinCompat {
-            name: "bls12_381_fp2_to_g2".to_string(),
-            pricing: PricingCompat::Single(Pricing::Linear(Linear {
-                base: 23_800,
-                word: 0,
-            })),
-            activate_at: None,
-        })
-        .try_into()
-        .unwrap(),
-    );
-
-    builtins
-}
-
 /// Runs a test in a separate thread with a specified stack size.
 ///
 /// # Panics
@@ -632,424 +147,15 @@ pub fn test(test_config: TestConfig, name: String, test: StateTestCase) -> TestE
     child.join().unwrap()
 }
 
-/// Validate EIP-3607 - empty create caller
-fn assert_empty_create_caller(expect_exception: Option<&String>, name: &str) {
-    let exception = expect_exception.expect("expected evm-json-test exception");
-    let check_exception =
-        exception == "SenderNotEOA" || exception == "TransactionException.SENDER_NOT_EOA";
-    assert!(
-        check_exception,
-        "expected EmptyCaller exception for test: {name}: {expect_exception:?}"
-    );
-}
-
-/// Check call expected exception
-fn assert_call_exit_exception(expect_exception: Option<&String>, name: &str) {
-    assert!(
-        expect_exception.is_none(),
-        "unexpected call exception: {expect_exception:?} for test: {name}"
-    );
-}
-
-/// Check Exit Reason of EVM execution
-fn check_create_exit_reason(
-    reason: &ExitReason,
-    expect_exception: Option<&String>,
-    name: &str,
-) -> bool {
-    match reason {
-        ExitReason::Error(err) => {
-            if let Some(exception) = expect_exception {
-                match err {
-                    ExitError::CreateContractLimit => {
-                        let check_result = exception == "TR_InitCodeLimitExceeded"
-                            || exception == "TransactionException.INITCODE_SIZE_EXCEEDED";
-                        assert!(
-                            check_result,
-                            "unexpected exception {exception:?} for CreateContractLimit error for test: {name}"
-                        );
-                        return true;
-                    }
-                    ExitError::MaxNonce => {
-                        let check_result = exception == "TR_NonceHasMaxValue"
-                            || exception == "TransactionException.NONCE_IS_MAX";
-                        assert!(check_result,
-                                "unexpected exception {exception:?} for MaxNonce error for test: {name}"
-                        );
-                        return true;
-                    }
-                    ExitError::OutOfGas => {
-                        let check_result =
-                            exception == "TransactionException.INTRINSIC_GAS_TOO_LOW";
-                        assert!(check_result,
-                                "unexpected exception {exception:?} for OutOfGas error for test: {name}"
-                        );
-                        return true;
-                    }
-                    _ => {
-                        panic!(
-                            "unexpected error: {err:?} for exception: {exception} for test: {name}"
-                        )
-                    }
-                }
-            } else {
-                return false;
-            }
-        }
-        ExitReason::Fatal(err) => {
-            panic!("Unexpected error: {err:?}")
-        }
-        _ => {
-            assert!(
-                expect_exception.is_none(),
-                "Unexpected json-test error: {expect_exception:?}"
-            );
-        }
-    }
-    false
-}
-
-/// Assert vicinity validation to ensure that test os expected validation error
 #[allow(clippy::cognitive_complexity, clippy::too_many_lines)]
-fn assert_vicinity_validation(
-    reason: &InvalidTxReason,
-    states: &[PostStateResult],
-    spec: &ForkSpec,
-    name: &str,
-    file_name: &Arc<PathBuf>,
-) {
-    match *spec {
-        ForkSpec::Istanbul | ForkSpec::Berlin => match reason {
-            InvalidTxReason::GasPriseEip1559 => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!(
-                            "expected error message for test: [{spec:?}] {name}:{i}\n{file_name:?}"
-                        )
-                    });
-
-                    let is_checked =
-                        expected == "TR_TypeNotSupported" || expected == "TR_TypeNotSupportedBlob";
-                    assert!(
-                        is_checked,
-                        "unexpected error message {expected:?} for: [{spec:?}] {name}:{i}\n{file_name:?}",
-                    );
-                }
-            }
-            _ => panic!("Unexpected validation reason: {reason:?} [{name}]\n{file_name:?}"),
-        },
-        ForkSpec::London => match reason {
-            InvalidTxReason::PriorityFeeTooLarge => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-                    let is_checked = expected == "tipTooHigh" || expected == "TR_TipGtFeeCap";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                        );
-                }
-            }
-            InvalidTxReason::GasPriceLessThenBlockBaseFee => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-                    let is_checked =
-                        expected == "lowFeeCap" || expected == "TR_FeeCapLessThanBlocks";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                        );
-                }
-            }
-            _ => {
-                panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}")
-            }
-        },
-        ForkSpec::Paris => match reason {
-            InvalidTxReason::PriorityFeeTooLarge => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-                    let is_checked = expected == "TR_TipGtFeeCap";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                        );
-                }
-            }
-            InvalidTxReason::GasPriceLessThenBlockBaseFee => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-                    let is_checked = expected == "TR_FeeCapLessThanBlocks";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                        );
-                }
-            }
-            _ => {
-                panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}")
-            }
-        },
-        ForkSpec::Shanghai => match reason {
-            InvalidTxReason::PriorityFeeTooLarge => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-                    let is_checked = expected == "TR_TipGtFeeCap";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}\n{file_name:?}",
-                        );
-                }
-            }
-            InvalidTxReason::GasPriceLessThenBlockBaseFee => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-
-                    let is_checked = expected == "TR_FeeCapLessThanBlocks";
-                    assert!(
-                            is_checked,
-                            "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                        );
-                }
-            }
-            _ => {
-                panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}")
-            }
-        },
-        ForkSpec::Cancun => match reason {
-            InvalidTxReason::PriorityFeeTooLarge => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-
-                    let is_checked = expected == "TR_TipGtFeeCap"
-                        || expected == "TransactionException.PRIORITY_GREATER_THAN_MAX_FEE_PER_GAS";
-                    assert!(
-                        is_checked,
-                        "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                    );
-                }
-            }
-            InvalidTxReason::GasPriceLessThenBlockBaseFee => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-
-                    let is_checked = expected == "TR_FeeCapLessThanBlocks"
-                        || expected == "TransactionException.INSUFFICIENT_MAX_FEE_PER_GAS";
-                    assert!(
-                        is_checked,
-                        "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                    );
-                }
-            }
-            _ => {
-                panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}")
-            }
-        },
-        ForkSpec::Prague => match reason {
-            InvalidTxReason::PriorityFeeTooLarge => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}")
-                    });
-
-                    let is_checked =
-                        expected == "TransactionException.PRIORITY_GREATER_THAN_MAX_FEE_PER_GAS";
-                    assert!(
-                        is_checked,
-                        "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                    );
-                }
-            }
-
-            InvalidTxReason::GasPriceLessThenBlockBaseFee => {
-                for (i, state) in states.iter().enumerate() {
-                    let expected = state.expect_exception.as_deref().unwrap_or_else(|| {
-                        panic!("expected error message for test: {reason:?} [{spec:?}] {name}:{i})\n{file_name:?}")
-                    });
-                    let is_checked = expected == "TR_FeeCapLessThanBlocks"
-                        || expected == "TransactionException.INSUFFICIENT_MAX_FEE_PER_GAS";
-                    assert!(
-                        is_checked,
-                        "unexpected error message {expected:?} for: {reason:?} [{spec:?}] {name}:{i}\n{file_name:?}",
-                    );
-                }
-            }
-            _ => {
-                panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}")
-            }
-        },
-        _ => panic!("Unexpected validation reason: {reason:?} [{spec:?}] {name}\n{file_name:?}"),
-    }
-}
-
-/// Check Exit Reason of EVM execution
-#[allow(clippy::too_many_lines)]
-fn check_validate_exit_reason(
-    reason: &InvalidTxReason,
-    expect_exception: Option<&String>,
-    name: &str,
-    spec: &ForkSpec,
-) -> bool {
-    expect_exception.map_or_else(
-        || {
-            panic!("unexpected validation error reason: {reason:?} {name}");
-        },
-        |exception| {
-            match reason {
-                InvalidTxReason::OutOfFund => {
-                    let check_result = exception
-                        == "TransactionException.INSUFFICIENT_ACCOUNT_FUNDS"
-                        || exception == "TR_NoFunds"
-                        || exception == "TR_NoFundsX"
-                        || exception == "TransactionException.INSUFFICIENT_MAX_FEE_PER_BLOB_GAS"
-                        || exception == "TransactionException.INSUFFICIENT_ACCOUNT_FUNDS|TransactionException.GASLIMIT_PRICE_PRODUCT_OVERFLOW";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for OutOfFund for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::GasLimitReached => {
-                    let check_result = exception == "TR_GasLimitReached"
-                        || exception == "TransactionException.GAS_ALLOWANCE_EXCEEDED";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for GasLimitReached for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::IntrinsicGas => {
-                    let check_result = exception == "TR_NoFundsOrGas"
-                        || exception == "TR_IntrinsicGas"
-                        || exception == "TransactionException.INTRINSIC_GAS_TOO_LOW"
-                        || exception == "IntrinsicGas"
-                        || exception == "TransactionException.INSUFFICIENT_ACCOUNT_FUNDS|TransactionException.INTRINSIC_GAS_TOO_LOW";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for IntrinsicGas for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::BlobVersionNotSupported => {
-                    let check_result = exception
-                        == "TransactionException.TYPE_3_TX_INVALID_BLOB_VERSIONED_HASH"
-                        || exception == "TR_BLOBVERSION_INVALID";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for BlobVersionNotSupported for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::BlobCreateTransaction => {
-                    let check_result = exception == "TR_BLOBCREATE"
-                        || exception == "TransactionException.TYPE_3_TX_CONTRACT_CREATION";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for BlobCreateTransaction for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::BlobGasPriceGreaterThanMax => {
-                    let check_result =
-                        exception == "TransactionException.INSUFFICIENT_MAX_FEE_PER_BLOB_GAS";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for BlobGasPriceGreaterThanMax for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::TooManyBlobs => {
-                    let check_result = exception == "TR_BLOBLIST_OVERSIZE"
-                        || exception == "TransactionException.TYPE_3_TX_BLOB_COUNT_EXCEEDED"
-                        || exception == "TransactionException.TYPE_3_TX_MAX_BLOB_GAS_ALLOWANCE_EXCEEDED|TransactionException.TYPE_3_TX_BLOB_COUNT_EXCEEDED";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for TooManyBlobs for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::EmptyBlobs => {
-                    let check_result = exception == "TransactionException.TYPE_3_TX_ZERO_BLOBS"
-                        || exception == "TR_EMPTYBLOB";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for EmptyBlobs for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::MaxFeePerBlobGasNotSupported => {
-                    let check_result =
-                        exception == "TransactionException.TYPE_3_TX_PRE_FORK|TransactionException.TYPE_3_TX_ZERO_BLOBS";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for MaxFeePerBlobGasNotSupported for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::BlobVersionedHashesNotSupported => {
-                    let check_result = exception == "TransactionException.TYPE_3_TX_PRE_FORK"
-                        || exception == "TR_TypeNotSupportedBlob";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for BlobVersionedHashesNotSupported for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::InvalidAuthorizationChain => {
-                    let check_result = exception == "TransactionException.TYPE_4_INVALID_AUTHORIZATION_FORMAT";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for InvalidAuthorizationChain for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::InvalidAuthorizationSignature => {
-                    let check_result = exception == "TransactionException.TYPE_4_INVALID_AUTHORITY_SIGNATURE";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for InvalidAuthorizationSignature for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::AuthorizationListNotExist => {
-                    let check_result = exception == "TransactionException.TYPE_4_EMPTY_AUTHORIZATION_LIST" || exception == "TransactionException.TYPE_4_TX_CONTRACT_CREATION";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for AuthorizationListNotExist for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::CreateTransaction => {
-                    let check_result = exception == "TransactionException.TYPE_4_TX_CONTRACT_CREATION";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for CreateTransaction for test: [{spec:?}] {name}"
-                    );
-                }
-                InvalidTxReason::GasFloorMoreThanGasLimit => {
-                    let check_result = exception == "TransactionException.INTRINSIC_GAS_TOO_LOW";
-                    assert!(
-                        check_result,
-                        "unexpected exception {exception:?} for GasFloorMoreThanGasLimit for test: [{spec:?}] {name}"
-                    );
-                }
-                _ => {
-                    panic!(
-                        "unexpected exception {exception:?} for reason {reason:?} for test: [{spec:?}] {name}"
-                    );
-                }
-            }
-            true
-        },
-    )
-}
-
-#[allow(clippy::cognitive_complexity, clippy::too_many_lines)]
-fn test_run(_test_config: &TestConfig, _name: &str, test: &StateTestCase) -> TestExecutionResult {
-    let tests_result = TestExecutionResult::new();
-    let _test_tx = &test.transaction;
+const fn test_run(
+    _test_config: &TestConfig,
+    _name: &str,
+    _test: &StateTestCase,
+) -> TestExecutionResult {
+    TestExecutionResult::new()
+    // let tests_result = TestExecutionResult::new();
+    // let _test_tx = &test.transaction;
     /*
         for (spec, states) in &test.post_states {
             // Run tests for specific SPEC (Hard fork)
@@ -1290,7 +396,7 @@ fn test_run(_test_config: &TestConfig, _name: &str, test: &StateTestCase) -> Tes
                 let actual_fee = executor.fee(vicinity.effective_gas_price);
                 // Forks after London burn miner rewards and thus have different gas fee
                 // calculation (see EIP-1559)
-                let miner_reward = if spec.is_eth2() {
+                let miner_reward = if *spec > Spec::Berlin {
                     let coinbase_gas_price = vicinity
                         .effective_gas_price
                         .saturating_sub(vicinity.block_base_fee_per_gas);
@@ -1383,7 +489,7 @@ fn test_run(_test_config: &TestConfig, _name: &str, test: &StateTestCase) -> Tes
             }
         }
     */
-    tests_result
+    // tests_result
 }
 
 /// Denotes the type of transaction.
@@ -1402,9 +508,9 @@ pub enum TxType {
 }
 
 impl TxType {
-    /// Whether this is a legacy, access list, dynamic fee, etc transaction
-    // Taken from geth's core/types/transaction.go/UnmarshalBinary, but we only detect the transaction
-    // type rather than unmarshal the entire payload.
+    /// Whether this is a legacy, access list, dynamic fee, etc. transaction
+    /// Taken from geth's core/types/transaction.go/UnmarshalBinary, but we only detect the transaction
+    /// type rather than unmarshal the entire payload.
     #[must_use]
     pub const fn from_txbytes(txbytes: &[u8]) -> Self {
         match txbytes[0] {
