+++ evm-tests/jsontests/src/vm.rs
@@ -1,10 +1,9 @@
-use crate::config::{TestConfig, VerboseOutput};
-use crate::state::{TestExecutionResult, VerboseOutput};
-use crate::types::StateTestCase;
-use aurora_evm::backend::{ApplyBackend, MemoryAccount, MemoryBackend, MemoryVicinity};
+use crate::config::VerboseOutput;
+use crate::execution_results::TestExecutionResult;
+use crate::types::VmTestCase;
+use aurora_evm::backend::{ApplyBackend, MemoryBackend};
 use aurora_evm::executor::stack::{MemoryStackState, StackExecutor, StackSubstateMetadata};
 use aurora_evm::Config;
-use primitive_types::{H160, H256, U256};
 use serde::Deserialize;
 use std::collections::BTreeMap;
 use std::rc::Rc;
@@ -12,92 +11,12 @@ use std::rc::Rc;
 #[derive(Deserialize, Debug)]
 pub struct Test(ethjson::vm::Vm);
 
-impl Test {
-    fn unwrap_to_pre_state(&self) -> BTreeMap<H160, MemoryAccount> {
-        crate::utils::unwrap_to_state(&self.0.pre_state)
-    }
-
-    fn unwrap_to_vicinity(&self) -> MemoryVicinity {
-        let block_randomness = self.0.env.random.map(|r| {
-            // Convert between U256 and H256. U256 is in little-endian but since H256 is just
-            // a string-like byte array, it's big endian (MSB is the first element of the array).
-            //
-            // Byte order here is important because this opcode has the same value as DIFFICULTY
-            // (0x44), and so for older forks of Ethereum, the threshold value of 2^64 is used to
-            // distinguish between the two: if it's below, the value corresponds to the DIFFICULTY
-            // opcode, otherwise to the PREVRANDAO opcode.
-            H256(r.0.to_big_endian())
-        });
-
-        MemoryVicinity {
-            gas_price: self.0.transaction.gas_price.into(),
-            effective_gas_price: self.0.transaction.gas_price.into(),
-            origin: self.0.transaction.origin.into(),
-            block_hashes: Vec::new(),
-            block_number: self.0.env.number.into(),
-            block_coinbase: self.0.env.author.into(),
-            block_timestamp: self.0.env.timestamp.into(),
-            block_difficulty: self.0.env.difficulty.into(),
-            block_gas_limit: self.0.env.gas_limit.into(),
-            chain_id: U256::zero(),
-            block_base_fee_per_gas: self.0.transaction.gas_price.into(),
-            block_randomness,
-            blob_gas_price: None,
-            blob_hashes: Vec::new(),
-        }
-    }
-
-    fn unwrap_to_code(&self) -> Rc<Vec<u8>> {
-        Rc::new(self.0.transaction.code.clone().into())
-    }
-
-    fn unwrap_to_data(&self) -> Rc<Vec<u8>> {
-        Rc::new(self.0.transaction.data.clone().into())
-    }
-
-    fn unwrap_to_context(&self) -> aurora_evm::Context {
-        aurora_evm::Context {
-            address: self.0.transaction.address.into(),
-            caller: self.0.transaction.sender.into(),
-            apparent_value: self.0.transaction.value.into(),
-        }
-    }
-
-    fn unwrap_to_return_value(&self) -> Vec<u8> {
-        self.0.output.clone().unwrap().into()
-    }
-
-    fn unwrap_to_gas_limit(&self) -> u64 {
-        self.0.transaction.gas.into()
-    }
-
-    fn unwrap_to_post_gas(&self) -> u64 {
-        self.0.gas_left.unwrap().into()
-    }
-
-    fn check_valid_state(&self, b: &BTreeMap<H160, MemoryAccount>) -> bool {
-        let post_state = self.0.post_state.as_ref().unwrap();
-        match &post_state.0 {
-            ethjson::spec::HashOrMap::Map(m) => {
-                &m.iter()
-                    .map(|(k, v)| ((*k).into(), crate::utils::unwrap_to_account(v)))
-                    .collect::<BTreeMap<_, _>>()
-                    == b
-            }
-            ethjson::spec::HashOrMap::Hash(h) => {
-                let x = crate::utils::check_valid_hash(&(*h).into(), b);
-                !x.0
-            }
-        }
-    }
-}
-
 #[must_use]
-pub fn test(
-    verbose_output: &VerboseOutput,
-    name: &str,
-    test: &StateTestCase,
-) -> TestExecutionResult {
+pub fn test(verbose_output: &VerboseOutput, name: &str, test: &VmTestCase) -> TestExecutionResult {
+    if "suicide" != name {
+        return TestExecutionResult::new();
+    }
+    println!("-- {name}");
     let mut result = TestExecutionResult::new();
     let mut failed = false;
     result.total = 1;
@@ -106,27 +25,32 @@ pub fn test(
         crate::utils::flush();
     }
 
-    let original_state = test.unwrap_to_pre_state();
-    let vicinity = test.unwrap_to_vicinity();
+    print!("## test {test:#?}");
+    let original_state = test.pre_state.to_memory_accounts();
+    let vicinity = test.get_memory_vicinity();
+    print!("## original_state {original_state:#?}");
+    print!("## vicinity {vicinity:#?}");
     let config = Config::frontier();
     let mut backend = MemoryBackend::new(&vicinity, original_state);
-    let metadata = StackSubstateMetadata::new(test.unwrap_to_gas_limit(), &config);
+    let metadata = StackSubstateMetadata::new(test.get_gas_limit(), &config);
     let state = MemoryStackState::new(metadata, &backend);
     let precompile = BTreeMap::new();
     let mut executor = StackExecutor::new_with_precompiles(state, &config, &precompile);
 
-    let code = test.unwrap_to_code();
-    let data = test.unwrap_to_data();
-    let context = test.unwrap_to_context();
-    let mut runtime =
-        aurora_evm::Runtime::new(code, data, context, config.stack_limit, config.memory_limit);
+    let mut runtime = aurora_evm::Runtime::new(
+        Rc::new(test.transaction.code.clone()),
+        Rc::new(test.transaction.data.clone()),
+        test.transaction.get_context(),
+        config.stack_limit,
+        config.memory_limit,
+    );
 
     let reason = executor.execute(&mut runtime);
     let gas = executor.gas();
     let (values, logs) = executor.into_state().deconstruct();
     backend.apply(values, logs, false);
 
-    if test.0.output.is_none() {
+    if test.output.is_none() {
         if verbose_output.verbose {
             print!("{reason:?} ");
         }
@@ -137,19 +61,19 @@ pub fn test(
                 print!("[Failed: succeed for empty output: {reason:?}] ");
             }
         }
-        if !(test.0.post_state.is_none() && test.0.gas_left.is_none()) {
+        if !(test.post_state.is_none() && test.gas_left.is_none()) {
             failed = true;
             if verbose_output.verbose_failed {
                 print!("[Failed: not empty state and left gas for empty output: {reason:?}] ",);
             }
         }
     } else {
-        let expected_post_gas = test.unwrap_to_post_gas();
+        let expected_post_gas = test.get_gas_left();
         if verbose_output.verbose {
             print!("{reason:?} ");
         }
 
-        if runtime.machine().return_value() != test.unwrap_to_return_value() {
+        if runtime.machine().return_value() != test.get_output() {
             failed = true;
             if verbose_output.verbose_failed {
                 print!(
@@ -158,7 +82,7 @@ pub fn test(
                 );
             }
         }
-        if !test.check_valid_state(backend.state()) {
+        if !test.validate_state(backend.state()) {
             failed = true;
             if verbose_output.verbose_failed {
                 print!("[Failed: invalid state] ");
